pyproject.toml
[project]
name = "rag-style-finder"
version = "0.1.0"
description = "Multimodal RAG Fashion Analysis Application"
authors = [
    {name = "V Crawford", email = "jonathancrawford05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core ML and Computer Vision
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",
    "pillow>=10.0.0",
    
    # Vector Search and Similarity
    "faiss-cpu>=1.8.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "requests>=2.31.0",
    
    # LangChain and LLM Integration
    "langchain>=0.2.6",
    "langchain-community>=0.2.6",
    "langchain-ollama>=0.1.0",
    
    # Ollama Integration
    "ollama>=0.3.0",
    
    # Web Interface
    "gradio>=4.44.1",
    
    # Additional utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.poetry]
packages = [{include = "rag_style_finder", from = "src"}]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
